Extract slide-out worktree tests into a dedicated `test_slide_out_worktrees.py`
Move the two worktree-conflict slide-out tests out of `test_slide_out.py` into their own
`test_slide_out_worktrees.py`, mirroring the existing `test_status_worktrees.py` /
`test_traverse_worktrees.py` split. The dedicated module carries a single module-level
`pytestmark = skipif(get_git_version() < (2, 5))` instead of repeating the decorator per test,
and the now-unused worktree-only imports (`textwrap`, `AbsPath`, `add_worktree`, `get_git_version`)
are dropped from `test_slide_out.py`.

Also trims the redundant `slide-out`-refuses-upfront release-notes bullet.
