+++ RELEASE_NOTES.md
@@ -4,7 +4,6 @@
 
 - added: `git machete status` now annotates each branch checked out in a worktree with a `[<worktree>]` label (suggested by @mikechristiansenvae)
 - fixed: `git machete go`, `slide-out` and `{github,gitlab} checkout-{pr,mr}s` now fail with an actionable error when the target branch is already checked out in another worktree, rather than the cryptic `fatal: '<branch>' is already used by worktree at <path>` from `git checkout` (reported by @lsierant)
-- fixed: `git machete slide-out` now refuses upfront, with the branch layout file left intact, when any branch it would need to `git checkout` after the slide is already held by another worktree (previously it would mutate `.git/machete` and only then bail out, leaving the layout half-applied) (reported by @lsierant)
 
 ## New in git-machete 3.41.0
